various fixes
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@100687 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: bea35446758b62e2ebe0e5e240838722bbb6d0d6

+++ Online/Online/OnlineHistDB/OnlineHistDB/OnlineHistogram.h
@@ -1,4 +1,4 @@
-// $Header: /afs/cern.ch/project/cvs/reps/lhcb/Online/OnlineHistDB/OnlineHistDB/OnlineHistogram.h,v 1.34 2010-06-08 17:18:05 ggiacomo Exp $
+// $Header: /afs/cern.ch/project/cvs/reps/lhcb/Online/OnlineHistDB/OnlineHistDB/OnlineHistogram.h,v 1.35 2010-06-10 16:57:50 ggiacomo Exp $
 #ifndef ONLINEHISTOGRAM_H
 #define ONLINEHISTOGRAM_H 1
 /** @class  OnlineHistogram OnlineHistogram.h OnlineHistDB/OnlineHistogram.h
@@ -252,7 +252,7 @@ class  OnlineHistogram : public OnlineHistDBEnv
                       std::vector<float>* alarm,
                       std::vector<float>* inputs,
                       bool &mask,
-                      int &statusbits,
+                      long &statusbits,
                       float &minstat,
                       float &minstatfrac); 
   /// masks analysis with ID AnaID. Use Mask=false to unmask. Returns true on success
